requirements.txt
adjustText==0.8
aiofiles==23.2.1
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
astroid==3.0.1
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
black==23.10.0
blinker==1.8.2
Cartopy==0.22.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.0
contourpy==1.2.0
cryptography==41.0.4
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
ecdsa==0.18.0
et-xmlfile==1.1.0
executing==2.0.0
fastapi==0.103.2
fiona==1.9.5
Flask==3.0.3
fonttools==4.46.0
geographiclib==2.0
geopandas==0.14.1
geopy==2.4.1
h11==0.14.0
httptools==0.6.0
idna==3.4
ipykernel==6.27.1
ipython==8.16.1
isort==5.12.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
mysqlclient==2.2.5
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
passlib==1.7.4
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.18.0
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.16.1
pylint==3.0.1
pyparsing==3.1.1
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rsa==4.9
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
stack-data==0.6.3
starlette==0.27.0
tenacity==8.2.3
tomlkit==0.12.1
tornado==6.4
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth==0.2.8
websockets==11.0.3
Werkzeug==3.1.2
scikit-learn